# Add support for CompletionList.CommitCharacters in LSP.