@NTaylorMullen 2021-05-03
> Aside - is this PR meant to target main?

Nope, forgot to associate you with the other PR though! 😄 Grouping the CompletionList.Data & CompletionList.CommitCharacter PRs together so they can re-use some of the infra